atualização concluida

+++ src/main/java/br/com/alunoonline/api/service/AlunoService.java
@@ -1,7 +1,7 @@
-package br.com.alunoonline.api.service;
+package alunoonline.alunoonline.service;
 
-import br.com.alunoonline.api.model.Aluno;
-import br.com.alunoonline.api.repository.AlunoRepository;
+import alunoonline.alunoonline.model.Aluno;
+import alunoonline.alunoonline.repository.AlunoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -16,15 +16,12 @@ public class AlunoService {
     @Autowired
     AlunoRepository alunoRepository;
 
-    public void criarAluno(Aluno aluno) {
-
-        alunoRepository.save(aluno);
-
+    public Aluno criarAluno(Aluno aluno) {
+        return alunoRepository.save(aluno);
     }
 
     public List<Aluno> listarTodosAlunos() {
         return alunoRepository.findAll();
-
     }
 
     public Optional<Aluno> buscarAlunoPorId(Long id) {
@@ -32,35 +29,27 @@ public class AlunoService {
     }
 
     public void deletarAlunoPorId(Long id) {
-        alunoRepository.deleteById(id);
+        Optional<Aluno> aluno = alunoRepository.findById(id);
+        if (aluno.isPresent()) {
+            alunoRepository.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Aluno não encontrado");
+        }
     }
 
-    public void atualizarAlunoPorId(Long id, Aluno aluno){
-        //primeiro passo: ver se o aluno existe no BD
-        Optional<Aluno> alunoDoBancodeDados = buscarAlunoPorId(id);
-
-        // e se nao existir esse aluno?
-        if (alunoDoBancodeDados.isEmpty()){
+    public void atualizarAlunoPorId(Long id, Aluno aluno) {
+        Optional<Aluno> alunoDoBancoDeDados = buscarAlunoPorId(id);
+        if (alunoDoBancoDeDados.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,
-                    "Aluno nao encontrado no Banco de dados");
-
+                    "Aluno não encontrado no banco de dados");
         }
-        // se chegar aqui, siginfica que existe com esse id
-        // vou armazenar em uma variavel par edita-lo
 
-        Aluno alunoParaEditar = alunoDoBancodeDados.get();
-
-        //com esse aluno para ser editado acima, faço os sets necessarios para atualizar ps atributos dele
+        Aluno alunoParaEditar = alunoDoBancoDeDados.get();
 
         alunoParaEditar.setNome(aluno.getNome());
         alunoParaEditar.setCpf(aluno.getCpf());
         alunoParaEditar.setEmail(aluno.getEmail());
-
-        //com aluno total editado acima eu devolvo ele para atualizado para BD
-
+        alunoParaEditar.setId(aluno.getId());
         alunoRepository.save(alunoParaEditar);
-
     }
-
-
 }
